Renamed realtime-user-monitoring to RealUserMonitoring.ts
