+++ sdk/src/tests/RaygunClient.test.ts
@@ -1,7 +1,7 @@
 import { Breadcrumb, BreadcrumbOption, CustomData, RaygunStackFrame } from '../types';
 import { internalStackFrames, stackFramesWithAddress, fullStackFrames } from './fixture/errors';
 import { sendCachedReports, sendCrashReport } from '../transport';
-import { setupRealtimeUserMonitoring } from '../realtime-user-monitor';
+import { setupRealtimeUserMonitoring } from '../RealUserMonitoring';
 
 jest.mock('../realtime-user-monitor', () => ({
   setupRealtimeUserMonitoring: jest.fn()
